+++ lib/Dialect/ESI/runtime/cpp/lib/backends/Trace.cpp
@@ -207,7 +207,7 @@ public:
   }
 
 private:
-  void connectImpl(std::optional<size_t> bufferSize) override {
+  void connectImpl(std::optional<unsigned> bufferSize) override {
     assert(!dataPushThread.joinable() && "already connected");
     shutdown = false;
     dataPushThread = std::thread(&ReadTraceChannelPort::dataPushLoop, this);
